Move cosmosinserter and hiringmanager to its own folder
